+++ docs/conf.py
@@ -15,6 +15,7 @@
 import sys
 import os
 import shlex
+import datetime
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -52,7 +53,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'Molo'
-copyright = u'2017, Praekelt.org'
+copyright = u'{}, Praekelt.org'.format(datetime.date.today().year)
 author = u'Praekelt.org'
 
 # The version info for the project you're documenting, acts as replacement for